requirements.txt
langchain>=0.2,<0.3
langchain-core>=0.2,<0.3
langgraph>=0.2,<0.3
pydantic>=2.0,<3
PyYAML>=6.0
numpy>=1.26
scikit-learn>=1.4
joblib>=1.3
boto3>=1.34
pytest>=8.0
scipy>=1.11
